% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sha_base64.R
\name{convert_sha_base64}
\alias{convert_sha_base64}
\title{Convert sha256 to base 64 using system call to xxd}
\usage{
convert_sha_base64(sha256)
}
\arguments{
\item{sha256}{Character vector with sha256 sum to be converted.}
}
\value{
Charecter vector with the converted string.
}
\description{
Convert sha256 to base 64 using system call to xxd
}
\note{
At the moment this function only works correctly on unix like systems.
This is because the system call relies on the pipe which is intepreted
differently in Windows.
}
