% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_local_sha.R
\name{read_local_sha}
\alias{read_local_sha}
\title{Read .lst file with sha}
\usage{
read_local_sha(path)
}
\arguments{
\item{path}{}
}
\value{
A data frame with two columns. The sha256 as returned by \code{sha256sum}.
}
\description{
Read .lst file with sha
}
